+++ base/docs/basedocs.jl
@@ -2357,8 +2357,8 @@ kw"::"
 """
     Vararg{T,N}
 
-The last parameter of a tuple type [`Tuple`](@ref) can be the special type `Vararg`, which denotes any
-number of trailing elements. The type `Vararg{T,N}` corresponds to exactly `N` elements of type `T`.
+The last parameter of a tuple type [`Tuple`](@ref) can be the special value `Vararg`, which denotes any
+number of trailing elements. `Vararg{T,N}` corresponds to exactly `N` elements of type `T`. Finally
 `Vararg{T}` corresponds to zero or more elements of type `T`. `Vararg` tuple types are used to represent the
 arguments accepted by varargs methods (see the section on [Varargs Functions](@ref) in the manual.)